extract tiptap code into simple module
