+++ domain/src/lib.rs
@@ -1,2 +1,4 @@
 pub mod coaching_session;
 pub mod error;
+
+pub(crate) mod gateway;
